pyproject.toml
[tool.poetry]
name = "ros-ocp-test-automation"
version = "0.1.0"
description = ""
authors = ["Suraj Patil <patilsuraj767@gmail.com>"]
readme = "README.md"
packages = [{include = "ros_ocp_test_automation"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pytest = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
